requirements.txt
streamlit==1.28.0
python-docx==0.8.11
pandas==2.0.3
Pillow==10.0.0